refacte cod
