+++ src/view/index/film-card.js
