+++ src/view/index/profile-rating.js
